+++ scripts/test_kpp_to_micm.py
@@ -10,7 +10,7 @@ Usage:
     pytest test_kpp_to_micm.py --log-cli-level=DEBUG
 """
 
-import kpp_to_micm
+from rxn_arrhenius import parse_kpp_arrhenius
 
 def test_parse_kpp_arrhenius():
     """
@@ -22,17 +22,17 @@ def test_parse_kpp_arrhenius():
 
     kpp_A, kpp_B, kpp_C = 1.0e-12, 2000.0, -3.0
 
-    arr_dict = kpp_to_micm.parse_kpp_arrhenius(
+    arr_dict = parse_kpp_arrhenius(
         'ARR_ab(%.2e, %.2f)' % (kpp_A, kpp_B))
     assert arr_dict['A'] == kpp_A
     assert arr_dict['C'] == - kpp_B
 
-    arr_dict = kpp_to_micm.parse_kpp_arrhenius(
+    arr_dict = parse_kpp_arrhenius(
         'ARR_ac(%.2e, %.2f)' % (kpp_A, kpp_C))
     assert arr_dict['A'] == kpp_A
     assert arr_dict['B'] == kpp_C
 
-    arr_dict = kpp_to_micm.parse_kpp_arrhenius(
+    arr_dict = parse_kpp_arrhenius(
         'ARR_abc(%.2e, %.2f, %.2f)' % (kpp_A, kpp_B, kpp_C))
     assert arr_dict['A'] == kpp_A
     assert arr_dict['C'] == - kpp_B
